+++ packages/cinterop/src/nativeDarwinTest/kotlin/io/realm/kotlin/test/CinteropTest.kt
@@ -82,7 +82,7 @@ import kotlin.test.assertTrue
 class CinteropTest {
     @Test
     fun version() {
-        assertEquals("13.5.0", realm_get_library_version()!!.toKString())
+        assertEquals("13.9.0", realm_get_library_version()!!.toKString())
     }
 
     @Test