+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/ConfigurationImpl.kt
@@ -136,6 +136,11 @@ public open class ConfigurationImpl constructor(
         this.initialDataCallback = initialDataCallback
         this.inMemory = inMemory
 
+        // Configure logging during creation of a (Realm/Sync)Configuration to keep old behavior
+        // for configuring logging. This should be removed when `LogConfiguration` is removed.
+        io.realm.kotlin.log.RealmLog.logLevel = log.level
+        log.loggers.forEach { io.realm.kotlin.log.RealmLog.addLogger(it) }
+
         // We need to freeze `compactOnLaunchCallback` reference on initial thread for Kotlin Native
         val compactCallback = compactOnLaunchCallback?.let { callback ->
             object : io.realm.kotlin.internal.interop.CompactOnLaunchCallback {