+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/LiveRealm.kt
@@ -26,6 +26,7 @@ import kotlinx.atomicfu.AtomicRef
 import kotlinx.atomicfu.atomic
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
+import io.realm.kotlin.log.RealmLog
 
 /**
  * A live realm that can be updated and receive notifications on data and schema changes when
@@ -49,7 +50,7 @@ internal abstract class LiveRealm(
     private val realmChangeRegistration: NotificationToken
     private val schemaChangeRegistration: NotificationToken
 
-    internal val versionTracker = VersionTracker(this, owner.log)
+    internal val versionTracker = VersionTracker(this)
 
     override val realmReference: LiveRealmReference by lazy {
         val (dbPointer, _) = RealmInterop.realm_open(configuration.createNativeConfiguration(), dispatcher)
@@ -97,7 +98,7 @@ internal abstract class LiveRealm(
         return snapshotLock.withLock {
             _snapshot.value.also { snapshot ->
                 if (_closeSnapshotWhenAdvancing) {
-                    log.trace("${this@LiveRealm} ENABLE-TRACKING ${snapshot.version()}")
+                    RealmLog.trace("${this@LiveRealm} ENABLE-TRACKING ${snapshot.version()}")
                     _closeSnapshotWhenAdvancing = false
                 }
             }
@@ -123,7 +124,7 @@ internal abstract class LiveRealm(
                 return
             }
             if (_closeSnapshotWhenAdvancing) {
-                log.trace("${this@LiveRealm} CLOSE-UNTRACKED $version")
+                RealmLog.trace("${this@LiveRealm} CLOSE-UNTRACKED $version")
                 _snapshot.value.close()
             } else {
                 // TODO Split into track and clean up as we don't need to hold headLock while
@@ -131,7 +132,7 @@ internal abstract class LiveRealm(
                 versionTracker.trackAndCloseExpiredReferences(_snapshot.value)
             }
             _snapshot.value = realmReference.snapshot(owner)
-            log.trace("${this@LiveRealm} ADVANCING $version -> ${_snapshot.value.version()}")
+            RealmLog.trace("${this@LiveRealm} ADVANCING $version -> ${_snapshot.value.version()}")
             _closeSnapshotWhenAdvancing = true
         }
     }
@@ -154,7 +155,7 @@ internal abstract class LiveRealm(
         realmReference.close()
         // Close current reference
         _snapshot.value?.let {
-            log.trace("$this CLOSE-ACTIVE ${it.version()}")
+            RealmLog.trace("$this CLOSE-ACTIVE ${it.version()}")
             it.close()
         }
         // Close all intermediate references