+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/log/LogLevel.kt
@@ -24,25 +24,4 @@ public enum class LogLevel(public val priority: Int) {
     ERROR(5),
     WTF(6),
     NONE(7);
-
-    public companion object {
-        /**
-         * Converts a Core log level value to a library log level value. Values that represent the
-         * same level from the library perspective are folded together.
-         *
-         * For internal use only.
-         */
-        public fun fromCoreLogLevel(coreLogLevel: CoreLogLevel): LogLevel = when (coreLogLevel) {
-            CoreLogLevel.RLM_LOG_LEVEL_ALL,
-            CoreLogLevel.RLM_LOG_LEVEL_TRACE -> TRACE
-            CoreLogLevel.RLM_LOG_LEVEL_DEBUG -> DEBUG
-            CoreLogLevel.RLM_LOG_LEVEL_DETAIL,
-            CoreLogLevel.RLM_LOG_LEVEL_INFO -> INFO
-            CoreLogLevel.RLM_LOG_LEVEL_WARNING -> WARN
-            CoreLogLevel.RLM_LOG_LEVEL_ERROR -> ERROR
-            CoreLogLevel.RLM_LOG_LEVEL_FATAL -> WTF
-            CoreLogLevel.RLM_LOG_LEVEL_OFF -> NONE
-            else -> throw IllegalArgumentException("Invalid core log level: $coreLogLevel")
-        }
-    }
 }